Change pre class to def for definitions.
This removes a bunch of parsing errors at the cost of minor
style changes to the blocks in the HTML. The errors this
fixes look like this:

LINE <number>: The elementdef for '???' is missing a 'Name' line.
LINE <number>: The elementdef for '???' is missing a 'Categories' line.
LINE <number>: The elementdef for '???' is missing a 'Contexts' line.
LINE <number>: The elementdef for '???' is missing a 'Content model' line.
LINE <number>: The elementdef for '???' is missing a 'Attributes' line.
LINE <number>: The elementdef for '???' is missing a 'Dom interfaces' line.